[WIP] Add an UnorderedFoldable that we can give instances for Set/Map/&c.
`Set`, e.g., is not a lawful `Foldable`, since iteration order
between multiple sets can be observed via `toList`. If we promise
never to observe the order in which elements are folded, by
strengthening the `Monoid` required by `foldMap` to a `CommutativeMonoid`,
we can have a broader typeclass `UnorderedFoldable` that contains
such unordered collections.

Based entirely off a gitter comment.