Reduce redundancy in Semigroup and Eq test names
Previously if you ran the `CommutativeSemigroup` tests on a type (`Mean`
in this example) you would see something like:

```
MeanTests:
- Mean.commutativeSemigroup.commutative
- Mean.commutativeSemigroup.semigroup associative
- Mean.commutativeSemigroup.semigroup combineAllOption
- Mean.commutativeSemigroup.semigroup repeat1
- Mean.commutativeSemigroup.semigroup repeat2
```

The extra `semigroup` in the name seems redundant to me, and as far as I
can tell, it's inconsistent with tests for other type classes. `Eq` had
a similar issue.