Be more precise about left vs right in Foldable doc
With the previous wording, a reader might interpret the `Foldable` docs
as suggesting that `foldLeft` traverses the structure from the left
while `foldRight` traverses the structure from the right. In my
experience, this is a common misconception. In fact, `foldRight` can
terminate on infinite structures while `foldLeft` can't, which doesn't
make sense if you are thinking about the two as traversal order. I've
tried to be more precise about `Left` and `Right` referring to the
associativity of the operations. I'm not sure whether my `List(1, 2, 3)`
example is very good or not -- open to feedback