Add FunctorFilter and TraverseFilter (#2405)
* Add FunctorEmpty and TraverseEmpty

* Add TraverseEmpty instance for Chain

* Make TraverseEmpty[Chain] serializable

I think that the issue here is that it was referencing an instance-level
`catsDataInstancesForChain` from an abstract class. By changing it to
reference `Chain.catsDataInstancesForChain`, it is a reference to a
static member (and therefore doesn't actually need to be serialized).

Take my explanation with a grain of salt -- like everyone else on the
planet, I don't actually understand Java serialization. But at the end
of the day it works :)

* Remove conversion to mainline classes

* Add traverseFilter <-> traverse consistency law

* Make Functor override final

* Rename Functor and Traverse Empty to Filter

* Add Nested FunctorFilter instance

* Address feedback