Documentation: add a nomenclature page.
This commit introduces a new page to the documentation,
showing the name and simplified type signatures of the methods
in the type classes and data types of the cats library.

Whereas other pages of the documentation try to explain one
type-class at a time in a manner that is amenable to beginners;
whereas the Scaladocs are good only for those who remember the
name of the type-class and want to recall what they do.

ScalaDocs, in particular, suffer from a limitation: they are
structured following the classes-packages-objects structure
of the Scala codeHowever, practitioners using a type-class
or a data type may be interested on browsing easyily through
different methods that may be defined in separate "modules".