Add awk script to parse test durations
This is something that I have recreated a couple of times when I've been
looking into builds timing out, so I wanted to save it somewhere in case
I want to use it again. I figured I might as well submit a PR in case
other people might find this to be useful. But I can understand if we
don't want to litter the Cats code base with one-off scripts.