Introduce Clay: light-weight Kleisli abstraction.
The Kleisli class is one of the well-known data types of Cats. A
Kleisli wraps a function (A => F[B]) into a.
By exposing the higher-kind parameter `F[_]` in the type system,
it makes it possible to define type-class instances and operations
that we could not define for functions.
This is an example of a type rotation: we re-arrange the types in the
nested type `=>[A, F[B]]` to make a plain `Kleisli[F, A, B]`.

However, one key problem of Kleisli is its performance. Because a
Kleisli is an outer object `Kleisli` that wraps an inner object
(a function), almost every method of the Kleisli class, and every
combinator of Kleisli objects, needs to allocate two objects:
first the new inner function, and then the wrapper.
This means that Kleisli-based libraries and applications are
paying a performance penalty to use this abstraction.

To remedy this problem, we introduce a lightweight version of Kleisli,
the trait `Clay[F[_], A, B]`, which is a trait with a single abstract
method `def apply(a: A): F[B]`. This has some advantages:

- This makes it almost similar to a function of A to an F[B], but
  without the methods of `Function1` that may get in the way.
- In Scala 2.12 and later we can use the Single-Abstract-Method
  syntax to define new Clay instances.

For the sake of feature parity, we implement in Clay all the methods
of Kleisli, as well as the type-class instances.