Don't create the hydra.license if $HydraLicense is empty
The sbt-hydra checks if a hydra.license file exists to decide whether it should
enable itself. Therefore, if the $HydraLicense is empty, it's important not to
create the hydra.license file.

The reason why the $HydraLicense can be empty is because its content is a
secret, hence it is available only to PRs created from within the repository.
Otherwise, the secret would be easily leaked.

This commit is a followup fix for 8069e09