Fix ReducibleLaws causing stack overflow by calling `Eval.now` early
The test `reduceRightConsistentWithReduceRightOption` currently calls
`value` on an `Eval` before the end of the computation, i.e. inside
the functions passed to `reduceRight` and `reduceRightOption`. This
can cause a stack overflow in otherwise lawful `Reducible` instances.

The test has been changed to instead use `map` on the `Eval` which is
the intended way of using `Eval` in methods like `reduceRight`.