Integrate cats-time into cats-kernel.
In [Issue 3766](https://github.com/typelevel/cats/issues/3766) it was
proposed to add java.time instances to cats itself, considering that
`java.time` is just as standard as UUID or the Scala library.

Within the companion objects of the typeclasses `Eq`, `Order`, and
`Hash`, we define new "mixin" traits that provide implementations
of each type-class based on the Any-hash, Any-eq, or Comparable.
These are only used to define simple instances of many typeclasses.
With that, each class in `java.time` has an instance of the kernel
typeclasses written in two lines, and they all fit in one file.

We also add instances of Show in the `core` module, which are
just based on the `toString` method.

- Duration Test: add a generator for durations