Integrate cats-time into cats-kernel.
The `cats-time` microlibrary publishes instances of the typeclasses of
cats.kernel and cats for the classes in the java.time package.
In [Issue 3766](https://github.com/typelevel/cats/issues/3766) it was
proposed to integrate this library into cats itself, considering that
`java.time` is just as standard as UUID or the Scala library.

Within the companion objects of the typeclasses `Eq`, `Order`, and
`Hash`, we define new "mixin" traits that provide implementations
of each type-class based on the Any-hash, Any-eq, or Comparable.
These are only used to define simple instances of many typeclasses.
With that, each class in `java.time` has an instance of the kernel
typeclasses written in two lines, and they all fit in one file.