Scala-3: Use Opaque Types.
Scala-3 comes with its own zero-cost Newtype implementation, as Opaque
Types. This is an improvement over existing case classes, or Newtype
encodings we are using.

In this commit, we bifurcate the `IdT` wrapper in two implementations:
the existing one goes to the Scala-2.x branch, a new one with newtype
is used on the new