Deprecate ifA.
The ifA method has been discussed as dangerous and misleading, because
executing the effects of _both_ branches is hardly ever desirable.
Proposed in Issue https://github.com/typelevel/cats/issues/3889

Instead, users of this method should use either the `ifM` method of
the `FlatMap` class (when each branch is an effect) or the `ifF`
method of the `Functor` class (where each branch is a value).