Deprecate OneAnd class.
The class `cats.data.OneAnd` represetns a pair of one element and
a collection of items. The main uses for this class, lists and lazy
lists (fka streams), have their own specialised class.
In issue https://github.com/typelevel/cats/issues/2089 it was
suggested that this class is not really useful and should be removed.

This commit marks the class, and the major methods that create
instances involving it, as deprecated.