Update Readme with badge for Scala support, fix other images
This changes the Scaladex badge in the README to show which versions of
Cats support what versions of Scala. I've chosen the variant which also
indicates Scala.js support, as Cats is being cross-published for exactly
the same Scala versions for the JVM and Scala.js. At the time of writing,
the badge says:

```
cats-core-2.6.1 (Scala 3.x, 2.13, 2.12 - Js 1.x), 2.0.0 (Scala 2.11 - Js 0.6), 1.2.0 (Scala 2.10 - Js 0.6)
```
[![cats-core Scala version support](https://index.scala-lang.org/typelevel/cats/cats-core/latest-by-scala-version.svg?targetType=Js)](https://index.scala-lang.org/typelevel/cats/cats-core) ![Continuous Integration](https://github.com/typelevel/cats/workflows/Continuous%20Integration/badge.svg)

More details on the badge format here: https://github.com/scalacenter/scaladex/pull/660


As well as this, I've fixed up a few broken images on the Cats Readme:

* Images around the 'Community Announcements' title, introduced with
  https://github.com/typelevel/cats/pull/2683, and served by
  https://placehold.it/ are unfortunately broken, so I've removed them.
* The Cats graphic at http://plastic-idolatry.com/erik/cats2.png used
  HTTP, so was considered insecure by modern browsers on the HTTPS web page
  https://typelevel.org/cats/ - leading to another broken image. I've
  replaced it with a GitHub-hosted copy.