Adds OrderedSet
OrderedSet is an newtype over SortedSet, but it enforces that all construction use an `Order`. The intent here is to guarantee that we are working with the coherent `Order` instance for some type `A`, allowing us to trivially construct lawful instances for `OrderedSet` which become non-trivial for `SortedSet` where we might be dealing with a non-coherent `Ordering`.