Add Wrapper Classes To Track Construction Of Order/Ordering
This commit adds two wrapper classes (private to kernel) which track the construction of Order from Ordering and vice versa. This permits us to check in some cases if a given set of Order/Ordering values are known to be of the same lineage, and thus consistent with each other.