restrict `traverse_` and friends to require `Unit`
When using `traverse_` or `sequence_` to evaluate some applicative
effect `G[B]` within the context of a foldable structure, any `B` value
is thrown away. Per the scaladoc for `traverse_`, these functions expect
that the `G[B]` is primarily a side effect or action.  The scala
convention for expressing this expectation is to require a `Unit`
parameter.

Teach `traverse_`, `sequence_`, and their parallel and nonempty
analogues to follow this convention by accepting only `G[Unit]` instead
of simply any `G[B]`.

Furthermore, update the implementations of `traverse_` and
`nonEmptyTraverse` to use `foldMapA` and `reduceMapA` respectively.
Since requiring Unit gives us a (trivial) `Monoid` for free, it turns
out that in the presence of such a monoid, the `***MapA` functions are
doing exactly the same thing as the existing implementations of
`traverse_` and `nonEmptyTraverse_`.